- Fix issues related to 64bit architectures where pointers are 64bit   and ints are 32bit.
(Logical change 1.266)

+++ src/lat_pagefault.c
@@ -23,7 +23,7 @@ typedef struct _state {
 	int clone;
 	char* file;
 	char* where;
-	int* pages;
+	size_t* pages;
 } state_t;
 
 void	initialize(iter_t iterations, void *cookie);
